requirements.txt
Django==5.1.3
gunicorn==21.2.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
whitenoise==6.7.0
dj-database-url==2.2.0
django-modeltranslation==0.19.17
Pillow==11.0.0
requests==2.32.3
